Added sparklineQuery.js and moved query
