+++ src/components/ProjectList/Queries/sparklineQuery.js
@@ -1,0 +1,14 @@
+import gql from 'graphql-tag';
+
+export const SPARKLINE = gql`
+  query Sparkline($name:String!) {
+  SparkyBoy(owner: "Lambda-School-Labs", name:$name){
+    id
+    message
+    additions
+    deletions
+    changedFiles
+    committedDate
+  }
+}
+`
